feat: add translations for banners
