+++ src/assets/lang/fr/banners.json
@@ -1,9 +1,18 @@
 {
   "valentinesBanner": {
-    "title": "L'amour sans limites!",
-    "subtitle": "90% de réduction sur notre plan 2 To avec le code VDAY",
-    "cta": "Obtenez l'offre",
-    "toast": "Copié sur le presse-papiers"
+    "label": "Économisez 69%",
+    "title": "Délices sucrés de la Saint-Valentin!",
+    "cta": "Choisissez un plan",
+    "guarantee": "Garantie de remboursement de 30 jours",
+    "lastCta": "*L'offre est valable pour les comptes gratuits ou les nouveaux clients",
+    "features": [
+      "Économisez sur les plans mensuels et annuels",
+      "Stockage cloud sûr et sécurisé",
+      "Open source et transparent",
+      "Transferts chiffrés de bout en bout",
+      "Aucun accès non autorisé",
+      "L'offre se termine le 25 février"
+    ]
   },
   "tryOutInternxtGeneralBanner": {
     "title": "Internxt",
