+++ src/assets/lang/ru/banners.json
@@ -26,10 +26,19 @@
     }
   },
   "valentinesBanner": {
-    "title": "Любовь без границ!",
-    "subtitle": "Получите 90% скидку на тарифный план 2 ТБ с кодом VDAY",
-    "cta": "Получить скидку",
-    "toast": "Скопировано в буфер обмена!"
+    "label": "Экономьте 69%",
+    "title": "Сладкие подарки ко Дню святого Валентина!",
+    "cta": "Выберите план",
+    "guarantee": "Гарантия возврата денег в течение 30 дней",
+    "lastCta": "*Предложение действует только для бесплатных аккаунтов или новых клиентов",
+    "features": [
+      "Получите 60% скидку на тарифный план 2 ТБ",
+      "Надежное и безопасное облачное хранилище",
+      "Резервное копирование файлов и фотографий",
+      "Сквозное шифрование данных",
+      "Отсутствие несанкционированного доступа",
+      "Предложение заканчивается 25 февраля"
+    ]
   },
   "tryOutInternxtGeneralBanner": {
     "title": "Познакомьтесь с Internxt",
