+++ src/assets/lang/zh/banners.json
@@ -29,10 +29,19 @@
     }
   },
   "valentinesBanner": {
-    "title": "Love without limits!",
-    "subtitle": "Get 90% off our 2TB plan with code VDAY",
-    "cta": "Get the deal",
-    "toast": "Copied to clipboard!"
+    "label": "省69%",
+    "title": "甜蜜的情人节礼物!",
+    "cta": "选择计划",
+    "guarantee": "30 天内退款保证",
+    "lastCta": "*此优惠仅适用于免费账户或新客户",
+    "features": [
+      "在月度和年度计划上节省",
+      "安全可靠的云存储",
+      "开源和透明",
+      "端到端加密传输",
+      "无未经授权访问",
+      "优惠截止日期为2月25日"
+    ]
   },
   "tryOutInternxtGeneralBanner": {
     "title": "Try Internxt",
